+++ packages/botonic-core/src/models/events/message/location.ts
@@ -1,7 +1,0 @@
-import { BotonicMessageEvent, MessageEventTypes } from './message-event'
-
-export interface LocationMessageEvent extends BotonicMessageEvent {
-  type: MessageEventTypes.LOCATION
-  lat: number
-  long: number
-}