+++ packages/botonic-core/src/models/events/message/replies.ts
@@ -1,8 +1,0 @@
-export interface Reply {
-  title: string
-  payload: string
-}
-
-export interface WithReplies {
-  replies?: Reply[]
-}