+++ packages/botonic-plugin-ai-agents/src/runner.ts
@@ -1,10 +1,15 @@
-import { Runner } from '@openai/agents'
+import {
+  InputGuardrailTripwireTriggered,
+  Runner,
+  RunToolCallItem,
+} from '@openai/agents'
 
 import {
   AgenticInputMessage,
   AgenticOutputMessage,
   AIAgent,
   Context,
+  RunResult,
 } from './types'
 
 export class AIAgentRunner {
@@ -18,7 +23,7 @@ export class AIAgentRunner {
     messages: AgenticInputMessage[],
     context: Context,
     maxRetries: number = 1
-  ): Promise<AgenticOutputMessage[]> {
+  ): Promise<RunResult> {
     return this.runWithRetry(messages, context, maxRetries, 1)
   }
 
@@ -27,18 +32,58 @@ export class AIAgentRunner {
     context: Context,
     maxRetries: number,
     attempt: number
-  ): Promise<AgenticOutputMessage[]> {
+  ): Promise<RunResult> {
     try {
       const runner = new Runner({
         modelSettings: { temperature: 0 },
       })
       const result = await runner.run(this.agent, messages, { context })
-      return result.finalOutput?.messages || []
+
+      const outputMessages = result.finalOutput?.messages || []
+      const hasExit =
+        outputMessages.length === 0 ||
+        outputMessages.some(message => message.type === 'exit')
+      const toolsExecuted = this.getExecutedNameTools(result)
+
+      return {
+        messages: hasExit
+          ? []
+          : (outputMessages.filter(
+              message => message.type !== 'exit'
+            ) as AgenticOutputMessage[]),
+        toolsExecuted,
+        exit: hasExit,
+        inputGuardrailTriggered: [],
+        outputGuardrailTriggered: [],
+      }
     } catch (error) {
+      if (error instanceof InputGuardrailTripwireTriggered) {
+        return {
+          messages: [],
+          toolsExecuted: [],
+          exit: true,
+          inputGuardrailTriggered: error.result.output.outputInfo,
+          outputGuardrailTriggered: [],
+        }
+      }
       if (attempt > maxRetries) {
         throw error
       }
       return this.runWithRetry(messages, context, maxRetries, attempt + 1)
     }
+  }
+
+  private getExecutedNameTools(result) {
+    return (
+      result.newItems
+        ?.filter(item => item instanceof RunToolCallItem)
+        .map((item: RunToolCallItem) => {
+          if (item.rawItem.type === 'function_call') {
+            return item.rawItem.name
+          }
+          return ''
+        })
+        .filter((toolName: string) => toolName !== '') || []
+    )
   }
 }