+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-ai-agent.tsx
@@ -1,16 +1,17 @@
+import { AgenticOutputMessage } from '@botonic/core'
 import { Button, Carousel, Text } from '@botonic/react'
 
-import { SOURCE_INFO_SEPARATOR } from '../constants'
-import { AgenticOutputMessage } from '../types'
+import { DO_NOTHING_PAYLOAD, SOURCE_INFO_SEPARATOR } from '../constants'
 import { ContentFieldsBase } from './content-fields-base'
 import { FlowElement } from './flow-element'
-import { HtAiAgentNode } from './hubtype-fields/ai-agent'
+import { HtAiAgentNode, HtInputGuardrailRule } from './hubtype-fields/ai-agent'
 
 export class FlowAiAgent extends ContentFieldsBase {
   public code: string = ''
   public name: string = ''
   public instructions: string = ''
   public activeTools?: { name: string }[]
+  public inputGuardrailRules: HtInputGuardrailRule[]
 
   public responses: AgenticOutputMessage[] = []
 
@@ -19,14 +20,15 @@ export class FlowAiAgent extends ContentFieldsBase {
     newAiAgent.name = component.content.name
     newAiAgent.instructions = component.content.instructions
     newAiAgent.activeTools = component.content.active_tools
-
+    newAiAgent.inputGuardrailRules =
+      component.content.input_guardrail_rules || []
     return newAiAgent
   }
 
   toBotonic(id: string): JSX.Element {
     return (
       <>
-        {this.responses.map(response => {
+        {this.responses.map((response: AgenticOutputMessage) => {
           if (response.type === 'text') {
             return <Text key={id}>{response.content.text}</Text>
           }
@@ -38,7 +40,7 @@ export class FlowAiAgent extends ContentFieldsBase {
                 {response.content.buttons.map((button, buttonIndex) => (
                   <Button
                     key={buttonIndex}
-                    payload={`do-nothing${SOURCE_INFO_SEPARATOR}${buttonIndex}`}
+                    payload={`${DO_NOTHING_PAYLOAD}${SOURCE_INFO_SEPARATOR}${buttonIndex}`}
                   >
                     {button}
                   </Button>