+++ packages/botonic-react/src/webchat/context/messages-reducer.ts
@@ -1,5 +1,4 @@
 import { SENDERS } from '../../index-types'
-import { isCustom } from '../../message-utils'
 import { WebchatAction } from './actions'
 import { WebchatState } from './types'
 
@@ -14,6 +13,8 @@ export const messagesReducer = (
       return addMessageComponent(state, action)
     case WebchatAction.UPDATE_MESSAGE:
       return updateMessageReducer(state, action)
+    case WebchatAction.UPDATE_CUSTOM_MESSAGE_PROPS:
+      return updateCustomMessagePropsReducer(state, action)
     case WebchatAction.UPDATE_REPLIES:
       return { ...state, replies: action.payload }
     case WebchatAction.REMOVE_REPLIES:
@@ -90,32 +91,29 @@ function updateMessageReducer(
   state: WebchatState,
   action: { type: WebchatAction; payload?: any }
 ) {
-  const msgIndex = state.messagesJSON.map(m => m.id).indexOf(action.payload.id)
+  const messageId = action.payload.id
+  const msgIndex = state.messagesJSON.map(m => m.id).indexOf(messageId)
   if (msgIndex > -1) {
     const msgComponent = state.messagesComponents[msgIndex]
-    let updatedMessageComponents = {}
+    let updatedMsgComponent = {}
     if (msgComponent) {
-      const updatedMsgComponent = {
+      updatedMsgComponent = {
         ...msgComponent,
         ...{
           props: { ...msgComponent.props, ack: action.payload.ack },
         },
       }
-
-      if (isCustom(action.payload)) {
-        // If the message is a custom message, update the json property to update props.
-        // If user close and open the chat the message will render again with the new props.
-        updatedMsgComponent.props.json = action.payload.data.json
-      }
-
-      updatedMessageComponents = {
-        messagesComponents: [
-          ...state.messagesComponents.slice(0, msgIndex),
-          { ...updatedMsgComponent },
-          ...state.messagesComponents.slice(msgIndex + 1),
-        ],
-      }
     }
+
+    const updatedMessagesComponents = msgComponent
+      ? getUpdatedMessagesComponents(state, msgIndex, updatedMsgComponent)
+      : state.messagesComponents
+
+    const messageJSON = state.messagesJSON.find(m => m.id === messageId)
+
+    const updatedMessagesJSON = messageJSON
+      ? getUpdatedMessagesJSON(state, msgIndex, action.payload)
+      : state.messagesJSON
 
     const numUnreadMessages = state.messagesComponents.filter(
       messageComponent => messageComponent.props.isUnread
@@ -123,12 +121,8 @@ function updateMessageReducer(
 
     return {
       ...state,
-      messagesJSON: [
-        ...state.messagesJSON.slice(0, msgIndex),
-        { ...action.payload },
-        ...state.messagesJSON.slice(msgIndex + 1),
-      ],
-      ...updatedMessageComponents,
+      messagesJSON: updatedMessagesJSON,
+      messagesComponents: updatedMessagesComponents,
       numUnreadMessages,
     }
   }
@@ -149,4 +143,77 @@ function addMessageReducer(
     ...state,
     messagesJSON: [...(state.messagesJSON || []), action.payload],
   }
+}
+
+function updateCustomMessagePropsReducer(
+  state: WebchatState,
+  action: { type: WebchatAction; payload?: any }
+) {
+  const { messageId, props } = action.payload
+
+  if (!messageId) {
+    return state
+  }
+
+  // Similar to updateMessageReducer but only for custom messages when update props
+  const msgIndex = state.messagesJSON.map(m => m.id).indexOf(messageId)
+  if (msgIndex > -1) {
+    const msgComponent = state.messagesComponents[msgIndex]
+    if (msgComponent) {
+      msgComponent.props = {
+        ...msgComponent.props,
+        ack: action.payload.ack,
+        ...props,
+      }
+    }
+
+    const updatedMessagesComponents = msgComponent
+      ? getUpdatedMessagesComponents(state, msgIndex, msgComponent)
+      : state.messagesComponents
+
+    const messageJSON = state.messagesJSON.find(m => m.id === messageId)
+    if (messageJSON) {
+      messageJSON.data = {
+        ...messageJSON.data,
+        ...props,
+      }
+    }
+
+    const updatedMessagesJSON = messageJSON
+      ? getUpdatedMessagesJSON(state, msgIndex, messageJSON)
+      : state.messagesJSON
+
+    return {
+      ...state,
+      messagesJSON: updatedMessagesJSON,
+      messagesComponents: updatedMessagesComponents,
+    }
+  }
+
+  return state
+}
+
+// Helper functions to update messagesComponents and messagesJSON
+function getUpdatedMessagesComponents(
+  state: WebchatState,
+  msgIndex: number,
+  updatedMessageComponent: any
+) {
+  return [
+    ...state.messagesComponents.slice(0, msgIndex),
+    { ...updatedMessageComponent },
+    ...state.messagesComponents.slice(msgIndex + 1),
+  ]
+}
+
+function getUpdatedMessagesJSON(
+  state: WebchatState,
+  msgIndex: number,
+  messageJSON: any
+) {
+  return [
+    ...state.messagesJSON.slice(0, msgIndex),
+    { ...messageJSON },
+    ...state.messagesJSON.slice(msgIndex + 1),
+  ]
 }