+++ packages/botonic-react/src/webchat/custom-messages/rating/index.tsx
@@ -1,5 +1,4 @@
 import { INPUT, InputType } from '@botonic/core'
-import merge from 'lodash.merge'
 import React, { useContext, useState } from 'react'
 import { ThemeContext } from 'styled-components'
 
@@ -14,56 +13,40 @@ interface CustomRatingMessageProps {
   messageText: string
   buttonText: string
   ratingType: RatingType
-  json: { messageId?: string; valueSent?: number }
+  id?: string // This id not exist in local development
+  valueSent?: number
 }
 
 const CustomRatingMessage: React.FC<CustomRatingMessageProps> = props => {
-  const { payloads, messageText, buttonText, ratingType } = props
-  const { webchatState, updateMessage, sendInput } = useContext(WebchatContext)
+  const { payloads, messageText, buttonText, ratingType, id, valueSent } = props
+  const { updateCustomMessageProps, sendInput } = useContext(WebchatContext)
 
   const theme = useContext(ThemeContext)
   const color = theme?.brand?.color ?? ''
 
-  const [ratingValue, setRatingValue] = useState(
-    props.json?.valueSent ? props.json.valueSent : -1
-  )
+  const [ratingValue, setRatingValue] = useState(valueSent ? valueSent : -1)
   const [showRating, setShowRating] = useState(true)
 
   const ratingChanged = (newRating: number) => {
     setRatingValue(newRating)
   }
 
-  const updateMessageJSON = (messageId?: string) => {
-    if (messageId) {
-      const messageToUpdate = webchatState.messagesJSON.filter(m => {
-        return m.id === messageId
-      })[0]
-      const messageInfo = {
-        data: {
-          json: {
-            valueSent: ratingValue,
-          },
-        },
-      }
-      const updatedMsg = merge(messageToUpdate, messageInfo)
-      updateMessage(updatedMsg)
-    }
-  }
-
   const handleButtonSend = () => {
-    if (ratingValue !== -1) {
-      setShowRating(false)
+    if (ratingValue === -1) return
 
-      const payload = payloads[ratingValue - 1]
-
-      updateMessageJSON(props.json?.messageId)
-
-      const input = {
-        type: INPUT.POSTBACK as InputType,
-        payload,
-      }
-      void sendInput(input)
+    const json = {
+      valueSent: ratingValue,
     }
+    updateCustomMessageProps(json, id)
+
+    setShowRating(false)
+    const payload = payloads[ratingValue - 1]
+
+    const input = {
+      type: INPUT.POSTBACK as InputType,
+      payload,
+    }
+    void sendInput(input)
   }
 
   const disabled = ratingValue === -1
@@ -77,10 +60,10 @@ const CustomRatingMessage: React.FC<CustomRatingMessageProps> = props => {
           onRatingChange={ratingChanged}
           ratingValue={ratingValue}
           ratingType={ratingType}
-          valueSent={props.json?.valueSent}
+          valueSent={valueSent}
         />
       </MessageBubble>
-      {!props.json?.valueSent && showRating && (
+      {!props?.valueSent && showRating && (
         <Button
           autodisable={true}
           disabled={disabled}